+++ src/Component/Event/Event.php
@@ -1,25 +1,15 @@
 <?php
+
 /**
- * This file is part of the Vection-Framework project.
- * Visit project at https://github.com/Vection-Framework/Vection
+ * This file is part of the Vection package.
  *
- * (c) Vection-Framework <vection@appsdock.de>
+ * (c) David M. Lung <vection@davidlung.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
 declare(strict_types=1);
-
-/*
- * This file is part of the Vection project.
- * Visit project at https://www.vection.de
- *
- * (c) Vection <project@vection.de>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
 
 namespace Vection\Component\Event;
 
@@ -29,10 +19,11 @@ use Vection\Contracts\Event\EventInterface;
  * Class Event
  *
  * @package Vection\Component\Event
+ *
+ * @author  David Lung <vection@davidlung.de>
  */
 abstract class Event implements EventInterface
 {
-
     /**
      * Stops the event propagation if set to true.
      * Use this to abort the event rising.
@@ -56,5 +47,4 @@ abstract class Event implements EventInterface
     {
         return $this->propagationStopped;
     }
-
 }