+++ src/Component/Event/Tests/Event/TestEvent.php
@@ -1,0 +1,26 @@
+<?php
+
+/**
+ * This file is part of the Vection package.
+ *
+ * (c) David M. Lung <vection@davidlung.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types = 1);
+
+namespace Vection\Component\Event\Tests\Event;
+
+/**
+ * Class TestEvent
+ *
+ * @package Vection\Component\Event\Tests\Event
+ *
+ * @author  David Lung <vection@davidlung.de>
+ */
+class TestEvent
+{
+
+}