+++ src/Contracts/Event/EventInterface.php
@@ -1,10 +1,9 @@
 <?php
 
 /**
- * This file is part of the Vection-Framework project.
- * Visit project at https://github.com/Vection-Framework/Vection
+ * This file is part of the Vection package.
  *
- * (c) Vection-Framework <vection@appsdock.de>
+ * (c) David M. Lung <vection@davidlung.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,12 +11,16 @@
 
 namespace Vection\Contracts\Event;
 
+use Psr\EventDispatcher\StoppableEventInterface;
+
 /**
  * Interface EventInterface
  *
  * @package Vection\Contracts\Event
+ *
+ * @author  David Lung <vection@davidlung.de>
  */
-interface EventInterface
+interface EventInterface extends StoppableEventInterface
 {
     /**
      * Stops the propagation of the event to further event handlers.