+++ src/Contracts/Event/EventListenerProviderInterface.php
@@ -1,0 +1,28 @@
+<?php
+
+/**
+ * This file is part of the Vection package.
+ *
+ * (c) David M. Lung <vection@davidlung.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types = 1);
+
+namespace Vection\Contracts\Event;
+
+use Psr\EventDispatcher\ListenerProviderInterface;
+
+/**
+ * Interface EventListenerProviderInterface
+ *
+ * @package Vection\Contracts\Event
+ *
+ * @author  David Lung <vection@davidlung.de>
+ */
+interface EventListenerProviderInterface extends ListenerProviderInterface
+{
+
+}