+++ src/Component/Event/CHANGELOG.md
@@ -1,1 +1,5 @@
-MAJOR.MINOR.PATCH (Y-M-D)
+# CHANGELOG for pre-release v0.*
+This changelog contains important changes done for this component.
+
+## v0.3.0 (2020-03-12)
+Complete rewrite event dispatcher with psr-14 support.